+++ CHANGELOG.md
@@ -8,6 +8,12 @@ Version 0.2.3.0
 
 <https://github.com/mstksg/backprop/releases/tag/v0.2.3.0>
 
+*   Explicit `collectVar` no longer needs a `ZeroFunc` for the container, and
+    so all versions of `collectVar` and functions that use it (`fmap`,
+    `liftA2`, `liftA3`, `traverse`, `mapAccumL`, `mapAccumR`) no longer require
+    `Backprop` or `Num` instances for the final returned container type.  This
+    enables a lot more flexibility in container types.  **Breaking change**
+    only for those using the *Explicit* interfaces.
 *   `BV` pattern synonym added to *Numeric.Backprop*, abstracting over
     application of `splitBV` and `joinBV`.